requirements.txt
fastapi
uvicorn[standard]
discord.py
groq
feedparser
python-dotenv
beautifulsoup4
httpx
lxml
